example c tests working

+++ src/shared/parameter_v2/config/examples/parameter.h
@@ -2,12 +2,14 @@
 // Parameter for C
 //
 // NOTE: Parameters for C will ALWAYS be immutable.
-// We don't use an API here
-// they have been created. Because:
+// We don't use an psuedo-class here to keep things simple
 //
-// - we don't want parameters to be a way of passing data between tasks, there are plenty of rtos
-//   message passing techniques we can use (queues, notifications, binary-sempahores, etc...)
-// - keeps the api simpler, only allow create, access and destroy of parameters/configs in c
+// - we don't want parameters to be a way of passing data between tasks, there are plenty
+// of rtos
+//   message passing techniques we can use (queues, notifications, binary-sempahores,
+//   etc...)
+// - keeps the api simpler, only allow create, access and destroy of parameters/configs in
+// c
 //
 // It is garunteed that don't change after `initAppDynamicParameters()` has been called
 //
@@ -20,26 +22,30 @@
 //
 // The compiler will NOT allow assignment to these structs without a struct initializer.
 //
-// However: nothing is preventing someone from "re-initializing" the struct leading to 
+// However: nothing is preventing someone from "re-initializing" the struct leading to
 // undefined behaviour. This can NOT be caught at compile time for obvious reasons.
 
-typedef struct BoolParameter {
+typedef struct BoolParameter
+{
     const bool value;
 } BoolParameter_t;
 
-typedef struct IntegerParameter {
+typedef struct IntegerParameter
+{
     const int value;
 } IntegerParameter_t;
 
-typedef struct UnsignedIntegerParameter {
-    const uint32_t value;
+typedef struct UnsignedIntegerParameter
+{
+    const unsigned value;
 } UnsignedIntegerParameter_t;
 
-typedef struct FloatParameter {
-    const int value;
+typedef struct FloatParameter
+{
+    const float value;
 } FloatParameter_t;
 
-typedef struct StringParameter {
+typedef struct StringParameter
+{
     const char* value;
 } StringParameter_t;
-
